[rename] 이름 변경 query => keyword
