+++ src/entities/reviews/index.ts
@@ -1,5 +1,6 @@
 export * from './model/types';
 export {getBestReviews} from './apis/api-service';
 export {default as ReviewList} from './ui/ReviewList';
-export {default as useSearchReviewsWithQuery} from './model/useSearchReviewsWithQuery';
+export {default as useSearchReviewsWithKeyword} from './model/useSearchReviewsWithQuery';
 export {default as ReviewArticle} from './ui/ReviewArticle';
+export {default as ReviewArticleLoading} from './ui/ReviewArticleLoading';
