--- src/features/search-review/index.ts
@@ -1,2 +1,0 @@
-export {default as SearchReviews} from './ui/SearchReviews';
-export {default as SearchReviewsLoading} from './ui/SearchReviewsLoading';
