+++ src/views/search/index.ts
@@ -1,2 +1,2 @@
 export {default as SearchPage} from './ui/SearchPage';
-export {default as SearchWithQueryPage} from './ui/SearchWithQueryPage';
+export {default as SearchWithKeywordPage} from './ui/SearchWithKeywordPage';
